rename: rename folders
removed all cache to enable case sensitivity in git
