+++ src/domain/MovementIndicator.js
